requirements.txt
annotated-types==0.7.0
anyio==4.9.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
distro==1.9.0
exceptiongroup==1.2.2
faiss-cpu==1.7.4
filelock==3.18.0
Flask==3.1.0
fsspec==2025.3.2
google-api-core==2.24.2
google-api-python-client==2.118.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.70.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.30.2
idna==3.10
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
joblib==1.5.0
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==1.26.4
openai==1.77.0
packaging==25.0
pillow==11.2.1
proto-plus==1.26.1
protobuf==6.30.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.4
pydantic_core==2.33.2
pyparsing==3.2.3
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.31.0
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==4.1.0
sniffio==1.3.1
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.7.0
tqdm==4.67.1
transformers==4.51.3
typing-inspection==0.4.0
typing_extensions==4.13.2
uritemplate==4.1.1
urllib3==2.4.0
Werkzeug==3.1.3
youtube-transcript-api==0.6.2
